Grading model rename files
