+++ server/src/database/seeders/001-dummy-data.ts
@@ -16,7 +16,7 @@ const users = readSql('users.sql');
 const courses = readSql('courses.sql');
 const courseRole = readSql('course_role.sql');
 const courseTranslation = readSql('course_translations.sql');
-const gradingModel = readSql('assessment_model.sql');
+const gradingModel = readSql('grading_model.sql');
 const attainment = readSql('attainment.sql');
 const attainmentGrade = readSql('attainment_grade.sql');
 const finalGrade = readSql('final_grade.sql');
